refactor: decouple component file
