+++ apps/web/app/(roles)/vagas/publique/page.tsx
@@ -8,16 +8,14 @@ import {
   LanguageSelect,
   TextInput,
 } from 'app/components/CustomFormField'
-import { FormRadioGroup } from 'app/components/FormRadioGroup'
 import { Button } from 'app/components/ui/button'
-import { FormMessage } from 'app/components/ui/form'
 import { useToast } from 'app/hooks/use-toast'
 import { SkillsField } from 'app/subscribers/profile/components/SkillsField'
-import { InputHTMLAttributes, useId, useRef } from 'react'
-import { FormProvider, useForm, useFormContext } from 'react-hook-form'
-import { Topics } from 'shared/src/enums/topics'
-import { useIntersectionObserver } from 'usehooks-ts'
-import { RolePreview, RolePreviewModal } from './RolePreview'
+import { InputHTMLAttributes } from 'react'
+import { FormProvider, useForm } from 'react-hook-form'
+import { RolePreviewModal } from './RolePreview'
+import { RolePreviewSection } from './RolePreviewSection'
+import { RoleTopic } from './RoleTopic'
 
 type FormFields = {
   name: keyof FormSchema
@@ -58,40 +56,6 @@ const fields: FormFields = [
     placeholder: 'Ex: Mínimo X anos de XP (a partir de $xx.xxx)',
   },
 ]
-
-const RolePreviewSection = () => {
-  const sectionRef = useRef<HTMLElement>(null)
-  const entry = useIntersectionObserver(sectionRef, {
-    threshold: 0.3,
-  })
-  const intersectionEntryIsNull = !entry
-  const isFloatSectionVisible = intersectionEntryIsNull
-    ? false
-    : !entry?.isIntersecting
-
-  return (
-    <>
-      {isFloatSectionVisible && (
-        <section className="fixed bottom-4 right-4 z-50 hidden w-[500px] rounded-md border border-solid border-gray-300 bg-white p-4 pt-6 shadow md:block">
-          <h2 className="scroll-m-20 border-b pb-2 text-xl font-semibold tracking-tight first:mt-0">
-            Pré visualização
-          </h2>
-          <section className="m-x-auto max-w-[425px]">
-            <RolePreview />
-          </section>
-        </section>
-      )}
-      <section ref={sectionRef} className="hidden bg-white pt-6 md:block">
-        <h2 className="scroll-m-20 border-b pb-2 text-xl font-semibold tracking-tight first:mt-0">
-          Pré visualização
-        </h2>
-        <section className="max-w-[425px]">
-          <RolePreview />
-        </section>
-      </section>
-    </>
-  )
-}
 
 export default function RolesCreate() {
   const form = useForm<FormSchema>({
@@ -192,45 +156,5 @@ export default function RolesCreate() {
         </section>
       </form>
     </FormProvider>
-  )
-}
-
-const RoleTopic = () => {
-  const { setValue, watch, formState } = useFormContext<FormSchema>()
-  const fieldId = 'topicsId'
-  const id = useId()
-
-  return (
-    <section className="flex flex-col gap-4 space-x-2">
-      <p className="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70">
-        Esta é uma vaga:
-      </p>
-      <FormRadioGroup
-        key={id}
-        options={[
-          {
-            label: 'Internacional',
-            value: Topics.INTERNATIONAL_VACANCIES.toString(),
-          },
-          {
-            label: 'Nacional',
-            value: Topics.NATIONAL_VACANCIES.toString(),
-          },
-        ]}
-        formKey={id}
-        selectedOption={
-          watch(fieldId) === Topics.INTERNATIONAL_VACANCIES.toString()
-            ? Topics.INTERNATIONAL_VACANCIES.toString()
-            : Topics.NATIONAL_VACANCIES.toString()
-        }
-        setSelectedOption={(option) => {
-          console.log(option)
-          setValue(fieldId, option)
-        }}
-      />
-      {formState.errors?.topicsId?.message && (
-        <FormMessage>{formState.errors?.topicsId.message}</FormMessage>
-      )}
-    </section>
   )
 }
